refactor(telegram): move flow-block machinery to flow.rs, verbatim (#471 phase 1)
First mechanical cut of the 7.2k-line handler.rs: the processing-log
flow (streaming state types, the three renderers, and the flow
operations: open/refresh/append/restick/freeze/fold/dedup) moves to a
dedicated flow.rs.

Pure move: code is byte-identical except visibility widened to
pub(crate) on the moved items and two handler helpers the flow calls
(format_inline; send_html_or_plain was already crate-visible), plus a
glob re-export from handler so every existing call site and test
import path keeps working unchanged. handler.rs: 7221 → 6318 lines;
flow.rs: 924.
